+++ CHANGELOG.md
@@ -8,6 +8,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.4.0-pre.1] - 2025-10-07
+
+### Changed
+
+- Static libraries were rebuilt with Xcode Version 16.4 (16F6).
+- Changed AR Foundation dependency version from 6.3.0 to 6.4.0-pre.1.
+
+### Deprecated
+
+- `ARKitMeshSubsystemExtensions` has been deprecated. Use `XRMeshSubsystem` instead.
+- `ARMeshClassification` has been deprecated. Use `XRMeshClassification` instead.
+
 ## [6.3.0] - 2025-09-29
 
 ### Changed