+++ Documentation~/whats-new.md
@@ -1,9 +1,12 @@
 ---
 uid: arkit-whats-new
 ---
-# What's new in version 6.3
+# What's new in version 6.4
 
-This release contains no significant changes to the Apple ARKit XR Plug-in.
+# Deprecated
+
+- `ARKitMeshSubsystemExtensions` has been deprecated in favor of a new cross-platform API. Use `XRMeshSubsystem` instead.
+- `ARMeshClassification` has been deprecated in favor of a new cross-platform API. Use `XRMeshClassification` instead.
 
 For a full list of changes in this version including backwards-compatible bugfixes, refer to the package [changelog](xref:arkit-changelog).